// apps/gateway/src/auth/dto/error-singup-codes.dto.ts
export enum StepOneErrors {
    ALREADY_IN_PROGRESS = 'ALREADY_IN_PROGRESS',
    EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
    TRANSPORTER_ERROR = 'TRANSPORTER_ERROR',
    SERVER_INTERNAL_ERROR = 'SERVER_INTERNAL_ERROR',
};

export enum StepThreeErrors {
    EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
    SERVER_INTERNAL_ERROR = 'SERVER_INTERNAL_ERROR',
};